requirements.txt
django
djangorestframework
google-generativeai
PyPDF2
python-dotenv